absorbed AllTracksModal into EditPlaylist
fixed PUT /api/playlists/:id and POST /api/playlists routes
