+++ react-app/src/modals/EditPlaylist.js
@@ -1,12 +1,11 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useHistory, useParams } from "react-router-dom";
+import { useHistory, useParams, NavLink } from "react-router-dom";
 import { editPlaylist } from "../store/playlist";
 import ModalForm from "../components/ModalForm/ModalForm";
 import ModalFormImage from "../components/ModalForm/ModalFormImage";
 import ModalFormInput from "../components/ModalForm/ModalFormInput";
 import ModalFormTextarea from "../components/ModalForm/ModalFormTextarea";
-import AllTracksModal from "../components/PlaylistFolders/EditPlaylistModal/AllTracksModal";
 import "../components/ModalForm/ModalForm.css";
 import "./ModalNav.css";
 // import "./AddEditPlaylist.css";
@@ -123,6 +122,48 @@ const BasicInfoForm = ({ setShowModal }) => {
       />
 
     </>
+  );
+};
+
+const AllTracksModal = ({ songArr }) => {
+  const { id } = useParams();
+  const dispatch = useDispatch();
+  const statePlaylists = useSelector(state => state.playlists);
+
+  const handleDelist = async (playlistId, songId) => {
+    const formData = new FormData();
+    const playlist = statePlaylists[playlistId];
+
+    formData.append("title", playlist?.title);
+    formData.append("description", playlist?.description || '');
+    formData.append(
+      "songs_order",
+      JSON.stringify(playlist?.songs_order.filter(id => id !== songId))
+    );
+
+    const res = await dispatch(editPlaylist(playlistId, formData));
+    if (res) return res;
+  };
+
+  return (
+    <div className="AEP_track_container">
+      <ul>
+        {songArr?.map((song) => (
+          <li key={song?.id} className="AEP_li flex-row">
+            <div className="flex-row">
+              <img src={song?.image_url} className="AEP_li_img" alt={song?.title} />
+              <NavLink to={`/songs/${song?.id}`}>{song?.title}</NavLink>
+            </div>
+            <p
+              onClick={handleDelist}
+              className="AEP_li_x"
+            >
+              &#10005;
+            </p>
+          </li>
+        ))}
+      </ul>
+    </div>
   );
 };
